% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{adjWeibull}
\alias{adjWeibull}
\title{Adjust three-parameter Weibull distribution to data}
\usage{
adjWeibull(
  x,
  amplitude = 10,
  shape,
  main = NULL,
  xlab = "Diameter class (cm)",
  ylab = "Relative frequency (\%)",
  mid = TRUE,
  line.col = "blue",
  bar.col = "grey80",
  cex.axis = 0.6,
  cex.lab = 0.8,
  las = 1,
  legendPos = "topright",
  ...
)
}
\arguments{
\item{x}{a numeric vector of observations.}

\item{amplitude}{a numeric value indicating the amplitude of classes.}

\item{shape}{optional description}
}
\description{
Adjust three-parameter Weibull distribution to data
}
\examples{
d <- rweibull(1000, shape = 2.5, scale = 15) + 80
res = adjWeibull(d, amplitude = 3, mid = FALSE, shape = 6,
                 main = "Weibull adjustment", line.col = "red",
                 legendPos = "right")
res
}
