% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{barkFactor}
\alias{barkFactor}
\title{Bark factor
The bark factor (k) is computed for trees in order to assess the importance
of the valuable wood in the overall volume of a tree (Husch et al., 1982):}
\usage{
barkFactor(dbh, thickness)
}
\arguments{
\item{dbh}{numeric, diameter over bark of the individual trees.}

\item{thickness}{numeric, bark thickness measured on individual trees.}
}
\description{
Bark factor
The bark factor (k) is computed for trees in order to assess the importance
of the valuable wood in the overall volume of a tree (Husch et al., 1982):
}
\references{
Husch, B., Miller, C., Beers, T., 1982. Forest mensuration.
Ronald Press Company, London, pp. 1 – 410.
}
