% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{basal_i}
\alias{basal_i}
\title{Individual basal area}
\usage{
basal_i(dbh)
}
\arguments{
\item{dbh}{numeric vector of diameter.}
}
\value{
Vector of individual basal area.
}
\description{
Individual basal area
}
\examples{
basal_i(dbh = 0.1)
}
