% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{basalContribution}
\alias{basalContribution}
\title{Basal area contribution}
\usage{
basalContribution(basal)
}
\arguments{
\item{basal}{numeric, basal area per species.}
}
\description{
The basal area contribution (in per cent) is defined as the part
of a given species trees in the overall basal area of all trees in an area.
}
