% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{blackman}
\alias{blackman}
\title{Index of Blackman}
\usage{
blackman(density)
}
\arguments{
\item{density}{numeric vector of the density.}
}
\value{
Index of Blackman.
}
\description{
Index of Blackman
}
