% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{circum}
\alias{circum}
\title{Circumference or perimeter}
\usage{
circum(dbh)
}
\arguments{
\item{dbh}{numeric vector of diameter}
}
\value{
A vector of circumference.
}
\description{
Computes circumference based on diameter.
}
\examples{
x = seq(1, 5, .4)
circum(x)
}
\seealso{
\code{\link{height}} for tree  height, \code{\link{dbh}} for diameter.
}
