% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{deg}
\alias{deg}
\alias{rad}
\alias{degree}
\alias{radian}
\title{Degree and Radian}
\usage{
deg(radian)

rad(degree)
}
\arguments{
\item{radian}{numeric, vector of radian values to be converted to degrees.}

\item{degree}{numeric, vector of degree values to be converted to radians.}
}
\value{
\code{deg} returns vector of degree values while
\code{rad} returns vector of radian values.
}
\description{
\code{deg} converts angle values from radians to degrees. \cr
\code{rad} converts angle values from degrees to radians.
}
\examples{
deg(pi / 2)
rad(180)

}
\seealso{
\code{\link{principal}}.
}
