% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{densityRegen}
\alias{densityRegen}
\title{Density of regeneration}
\usage{
densityRegen(data = NULL, plot = NULL, count, nbSubPlot, area)
}
\arguments{
\item{data}{an optional data frame, list, tibble or object coercible by
\code{\link{as.data.frame}} to a data frame containing the variables whose
names are given in \code{count} and \code{plot}.}

\item{plot}{an optional character, name of the variable containing the plot
identities. If \code{data} is missing, a vector providing the plot identities.}

\item{count}{character, name of the variable containing the counts: number
of stems (individuals). If \code{data} is missing, a numeric vector
providing the the counts: number of stems (individuals).}

\item{nbSubPlot}{numeric, number of subplots per plot.}

\item{area}{numeric, area of each subplot.}
}
\description{
Computes the density per plot of tree regeneration based on counts in subplots.
}
