% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{densityTree}
\alias{densityTree}
\title{Tree density}
\usage{
densityTree(number, area, overall = TRUE)
}
\arguments{
\item{number}{numeric vector of individual count in each plot.}

\item{area}{numeric, area of a plot.}

\item{overall}{logical. If \code{TRUE}, an overall density is
computed; if \code{FALSE}, density is computed for each plot.
Default is \code{TRUE}.}
}
\value{
Vector of density.
}
\description{
Density per plot.
}
\details{
If every plot have same area, \code{area} is a numeric otherwise
\code{area} is a vector of each plot area.
}
