% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{diversityresult}
\alias{diversityresult}
\title{Function modified from 'BiodiversityR' package to allow 'base' argument}
\usage{
diversityresult(
  x,
  y = NULL,
  factor = NULL,
  level = NULL,
  base = 2,
  index = c("Shannon", "Simpson", "inverseSimpson", "Logalpha", "Berger", "richness",
    "abundance", "Jevenness", "Eevenness", "jack1", "jack2", "chao", "boot"),
  method = c("pooled", "each site", "mean", "sd", "max", "jackknife"),
  sortit = FALSE,
  digits = 8
)
}
\description{
Function modified from 'BiodiversityR' package to allow 'base' argument
}
