% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{diversityvariables}
\alias{diversityvariables}
\title{Function modified from 'BiodiversityR' package to allow 'base' argument}
\usage{
diversityvariables(x, y, digits = 8, base = 2)
}
\description{
Function modified from 'BiodiversityR' package to allow 'base' argument
}
