% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{factorize}
\alias{factorize}
\alias{factorise}
\title{Making factor vectors}
\usage{
factorize(data, binary = FALSE)
}
\arguments{
\item{data}{data frame or tibble data set.}

\item{binary}{logical indicating if binary data should be considered as factor.
Default is \code{FALSE}.}
}
\value{
Data frame with all character vectors changed to factor vectors.
}
\description{
Changes character vectors of a data set to factor vectors.
}
\details{
When \code{binary = TRUE}, variables stored as numeric and which have
exactly two levels are changed to factor.
}
