% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{girard}
\alias{girard}
\title{Girard Form Class
Girard Form Class is a form quotient used to estimate taper.}
\usage{
girard(dbh, dbhIn)
}
\arguments{
\item{dbh}{numeric, diameter outside bark at breast height.}

\item{dbhIn}{numeric, diameter inside bark at the top of the first log}
}
\description{
Girard Form Class
Girard Form Class is a form quotient used to estimate taper.
}
\references{
Strimbu, B. (2021). Dendrometry Field Manual.
}
