% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\docType{data}
\name{Logging}
\alias{Logging}
\title{Tree metrics for logging}
\format{
Data frame with twenty five observations and eight variables:
\describe{
  \item{tree}{Tree name}
  \item{hauteur}{Stem lenght in meter (m).}
  \item{diametreMedian}{Tree median diameter in centimeter (cm).}
  \item{perimetreMedian}{Tree median circumference in centimeter (cm).}

  \item{diametreSection}{Tree diameter at the end in centimeter (cm).}
  \item{perimetreSection}{Tree circumference at the end in centimeter (cm).}

  \item{diametreBase}{Tree diameter at the base in centimeter (cm).}
  \item{perimetreBase}{Tree circumference at the base in centimeter (cm).}
}
}
\source{
Fake data simulated for tutorial purposes.
}
\usage{
data(Logging)
}
\description{
Data frame of 24 rows and 8 columns containing tree measures.
}
\examples{
#demo(volume)
}
\author{
Narcisse Yehouenou \email{narcisstar211@gmail.com}
}
\keyword{datasets}
