% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{makedata}
\alias{makedata}
\title{Make stand data}
\usage{
makedata(data, factor1 = "", factor2 = "", factor3 = "")
}
\arguments{
\item{data}{data frame containing optional factors \code{factor1,factor2,factor3}.}

\item{factor1, factor2, factor3}{optional variables of the data frame that define subsets to consider.}
}
\value{
A list of data.
}
\description{
Make data of stands according to defined \code{factor1,factor2,factor3}.
}
\examples{
# require(BiodiversityR)
# data(ifri, package = "BiodiversityR")
#a1=makedata(ifri, factor1 = "forest", factor2 = "plotID", factor3 = "species")
#a2=makedata(ifri, factor1 = "species")
#makedata(ifri, factor2 = "")
#identical(makedata(ifri), ifri)
}
