% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{ombrothermique}
\alias{ombrothermique}
\title{Climograph
Plots the climograph of a location. The climograph is a representation of basic climate of a location. It displays monthly average precipitation and monthly average temperature.}
\usage{
ombrothermique(prec, temp, month, title = "Climograph")
}
\arguments{
\item{prec}{description}

\item{temp}{description}

\item{month}{description}

\item{title}{description}
}
\description{
Climograph
Plots the climograph of a location. The climograph is a representation of basic climate of a location. It displays monthly average precipitation and monthly average temperature.
}
