% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{param}
\alias{param}
\title{Structural parameter}
\usage{
param(data, factor1 = "Commune", factor2 = "Releve", factor3 = "Species",
area = 0.15, DBH = "DBH", height = "Height", houppier = "houppier", group = TRUE)
}
\description{
Structural parameter
}
