% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{paramDendro}
\alias{paramDendro}
\title{param for general; To be deleted}
\usage{
paramDendro(
  data,
  Site = "Releve",
  area = 0.15,
  DBH = "DBH",
  height = "Height",
  houppier = "houppier",
  group = TRUE
)
}
\description{
param for general; To be deleted
}
