% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{print.angle}
\alias{print.angle}
\title{Print Angle}
\usage{
\method{print}{angle}(x, ...)
}
\arguments{
\item{x}{an angle object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Method to print angle and returns it invisibly.
}
