% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{print.slope}
\alias{print.slope}
\title{Print Slope}
\usage{
\method{print}{slope}(x, ...)
}
\arguments{
\item{x}{a slope object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Method to print slope and returns it invisibly.
}
