% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{rad}
\alias{rad}
\title{Degrees to radians}
\usage{
rad(degree)
}
\arguments{
\item{degree}{A numeric vector of degree values to be converted.}
}
\value{
A vector of radian values.
}
\description{
Converts angle values from degree to radian.
}
\examples{
rad(180)
}
\seealso{
\code{\link{deg}}, the reciprocal of \code{rad}.
}
