% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{raster2df}
\alias{raster2df}
\title{Create Data Frame from Raster}
\usage{
raster2df(
  biodata,
  ext = c(0, 4, 6, 14),
  rmdup = TRUE,
  plot = TRUE,
  which = "c"
)
}
\arguments{
\item{biodata}{Stack raster}

\item{ext}{Spatial extent ...}

\item{plot}{Logical. Should raters be plotted ?}

\item{which}{character, indicates data in which duplicates should be checked.
'c' for coordinates, 'd' for data without coordinates, 'b' for both.}

\item{smdup}{Logical. Should duplicates be removed ?}
}
\description{
Create a data frame from (stack) raster
}
