% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{shared}
\alias{shared}
\title{Shared elements or intersection}
\usage{
shared(x, y)
}
\arguments{
\item{x, y}{vectors which shared elements (intersection) are needed.}
}
\value{
A list whose first element is the length of shared elements
between two vectors and second element is a vector of shared elements.
}
\description{
Determine the number of shared elements by two vector
as well as the shared elements.
}
