% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{skewness}
\alias{skewness}
\title{Skewness coefficient}
\usage{
skewness(x)
}
\arguments{
\item{x}{numeric vector.
The skewness.}
}
\value{
The skewness coefficient.
}
\description{
Skewness coefficient
}
\examples{
data("Logging")
skewness(Logging$hauteur)
hist(Logging$hauteur,3)
}
