% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{sorensenAll}
\alias{sorensenAll}
\title{Sorensen index for entire study area}
\usage{
sorensenAll(data, species, env)
}
\arguments{
\item{data}{data frame}

\item{species}{a character specifying species column name.}

\item{env}{a character specifying environment column name.}
}
\value{
Matix of Sorensen index.
}
\description{
Sorensen index for entire study area
}
