% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{stacking}
\alias{stacking}
\title{Stack all vectors of a data frame or list}
\usage{
stacking(data)
}
\arguments{
\item{data}{data frame, tibble or list.}
}
\value{
A vector of all element of the argument \code{data}.
}
\description{
Stacking all columns of a data frame or vectors of a list into a single vector.
}
