\name{taperPlot}
\alias{taperPlot}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{ Plot Taper Profiles from Dendrometered Trees }

\description{ This routine will generate a lattice conditional plot for
  one or more trees in an list returned from
  \code{\link{SplineVolume}}. The plots can include a splined version of
  the profile if desired.
}

\usage{
taperPlot(ds,
          TreeNos = NA,
          diamKind = c("dib", "dob"),
          diamType = c("diameter", "radius", "xsection"),
          standUp = FALSE,
          ylab = NULL,
          xlab = NULL,
          rugReadings = c('SC','PC'), 
          method = "mono",
          maxTrees = 50,
          fileName = "",
          pointsize.text = 12,
          ...)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ds}{ A list object of class \dQuote{SplineVolume}, that has been
	generated by \code{\link{SplineVolume}}.
}

  \item{TreeNos}{ The tree numbers in the collection to be plotted. If
	\code{NA} or \code{NULL}, the entire collection of trees in the
	object will be plotted. Otherwise, this should be a vector of tree
	numbers. Any not found in the collection are ignred.
}

  \item{diamKind}{ One can choose to use diameter inside or outside
	bark, though they may both be the same if no bark thinckness
	information was taken in the dendrometry.
}

  \item{diamType}{ Once can also choose to use diameter, radius or
	cross-sectional area in the plots.
  }
  
  \item{standUp}{ \code{TRUE}: align the tree height vertically;
	\code{FALSE} align it horizontally (default).
}

  \item{ylab}{ The y-axis label for plotting, the default depends on
  \code{standUp} and \code{diamType}. Beyond this, you should pass your
  own labels to make sense of the graph (i.e., for inside or outside
  bark). See \code{\link{xyplot}} for more information. }

  \item{xlab}{ The x-axis label for plotting--see \code{ylab} for more
	information. }

  \item{rugReadings}{ A vector of character product codes indicating
	where in the sense of tree height readings you would like
	\sQuote{rug} marks to appear in the graphs. For example, the
	defaults \code{c('SC','PC')} are two codes the author used to
	signify crown base in sawtimber and pulpwood products. Use \code{NA}
	or \code{NULL} for no rug marks.}

  \item{method}{ The spline method to use in plotting if desired. The
	method must be one that is compatible with
	\code{\link{splinefun}}. Note that if no spline is desire, use
	\code{NA} or \code{NULL}.
}

  \item{maxTrees}{ Limits the maximum nmber of tree profiles in any one
	lattice plot. This is obviously resettable and is there simply as a
	way to alert you about perhaps plotting hundreds of trees when you
	did not want to. Override as desired.
}

  \item{fileName}{ The file name for the output file. Please see
	\code{\link{hardcopyLattice}} for more details. A blank string
	(default) signals no hard copy.
}

  \item{pointsize.text}{ The point size for the hard copy if desired. Please see
	\code{\link{hardcopyLattice}} for more details.
}

  \item{\dots}{ Arguments that are to be passed to \code{\link{xyplot}}.
}
}

\details{ The default produces one lattice condition panel for each
  tree. it is quite simple. Please note that you must pass it a legal
  list of S3 class \dQuote{SplineVolume} for it to work. There are few
  error checks other than this.
}

\value{ The function returns the Lattice plot object invisibly.
}

\author{
Jeffrey Gove
}



%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{xyplot}}
}

\examples{
#
# assume dendro is an object from SplineVolume, then, e.g.,...
#
\dontrun{
plt = taperPlot(dendro, type='l', grid=T, diamType='xsection', standUp=TRUE)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
