% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\docType{data}
\name{Tree}
\alias{Tree}
\title{Dendrometric measures on tree}
\format{
Data frame with ten observations and five variables:
\describe{
  \item{circum}{Tree circumference in centimeter (cm).}
  \item{dist}{Horizontal distance between the observer (person who measure angles) and the tree circumference in centimeter (cm).}
  \item{up}{Up angle measure in degree (°).}
  \item{down}{Down angle measure in degree (°).}
  \item{fut}{Bole angle measure in degree (°); Bole is where the first branch occurs on the trunk. This measure is usually useful for timber estimation on wood market.}
}
}
\source{
Fake data simulated for tutorial purposes.
}
\usage{
data(Tree)
}
\description{
Data frame of 10 rows and 5 columns containing tree measures.
}
\examples{
#demo(dendro)
}
\author{
Narcisse Yehouenou \email{narcisstar211@gmail.com}
}
\keyword{datasets}
