% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{deg}
\alias{deg}
\title{Radians to degrees}
\usage{
deg(radian)
}
\arguments{
\item{radian}{A vector of radian values to be converted.}
}
\value{
A vector of degree values.
}
\description{
Converts angle values from  radian to degree.
}
\examples{
deg(pi/2)
}
\seealso{
\code{\link{rad}}, the reciprocal of \code{deg}.
}
