% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{diameterMean}
\alias{diameterMean}
\title{Mean diameter}
\usage{
diameterMean(dbh, factor1 = "", factor2 = "", factor3 = "", data)
}
\arguments{
\item{dbh}{numeric vector of diameter.}

\item{factor1, factor2, factor3}{optional variables of the data frame that define subsets to consider.}

\item{data}{data frame containing optional factors \code{factor1,factor2,factor3}.}
}
\value{
Mean diameter.
}
\description{
Mean diameter of a forestry stand.
}
\examples{
set.seed(1)
diameter = rnorm(10, 100, 20)
diameterMean(dbh = diameter)
}
\seealso{
\code{\link{dbh}}, \code{\link{basal_i}}
}
