% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\name{factorize}
\alias{factorize}
\alias{factorise}
\title{Making factor vectors}
\usage{
factorize(data, binary = FALSE)
}
\arguments{
\item{data}{A data set containing ....}

\item{binary}{Logical indicating if binary data should be considered as factor}
}
\value{
Data frame with all character vectors changed to factor vectors
}
\description{
Changes character vectors to factor vectors
}
