% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrometry.R
\docType{data}
\name{Tree}
\alias{Tree}
\title{Dendrometric measures on tree}
\format{
Data frame with ten observations and five variables:
\describe{
  \item{circum}{Tree circumference in centimeter (cm).}
  \item{dist}{Horizontal distance between the person measuring angles and
  the tree (m).}
  \item{up}{Angle measured for the top part of the tree in degree (\\u00b0).
  It is used to calculate the total tree height.}
  \item{down}{Angle measured for the bottom part of the tree in degree (\\u00b0).}
  \item{fut}{Bole angle measure in degree (\\u00b0); Bole is where the first branch
  occurs on the trunk. It is used to calculate the merchantable tree height.}
}
}
\source{
Fake data simulated for tutorial purposes.
}
\usage{
data(Tree)
}
\description{
Data frame of 10 rows and 5 columns containing tree measures.
}
\examples{
# demo(dendro)
}
\author{
Narcisse Yehouenou \email{narcisstar211@gmail.com}
}
\keyword{datasets}
